+++ packages/emoji-picker/CHANGELOG.md
@@ -15,6 +15,8 @@
 
 - Added the ability to clear commonly used emoji. Can be customized with the `clearIcon` prop.
 - Added titles/messages for skin tones within the palette.
+- Added a `noPreview` prop that renders content when there is no preview.
+- Added a `noResults` prop that renders content when there is no search results.
 - Added a `stickyGroupHeader` prop to enable sticky group headers.
 - Added an `emojiLoading` class name. Can be used to style a loading state for each emoji.
 - Added an `emojisHeaderSticky` class name. Can be used to style stickied group headers.